+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/spherical/SphericalSurfaceView.java
@@ -37,9 +37,14 @@ import android.view.Display;
 import android.view.Surface;
 import android.view.WindowManager;
 import com.google.android.exoplayer2.C;
-import com.google.android.exoplayer2.ui.spherical.Mesh.EyeType;
+import com.google.android.exoplayer2.Format;
+import com.google.android.exoplayer2.ui.spherical.ProjectionRenderer.EyeType;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Util;
+import com.google.android.exoplayer2.video.VideoFrameMetadataListener;
+import com.google.android.exoplayer2.video.spherical.Projection;
+import com.google.android.exoplayer2.video.spherical.ProjectionDecoder;
+import java.util.Arrays;
 import javax.microedition.khronos.egl.EGLConfig;
 import javax.microedition.khronos.opengles.GL10;
 
@@ -54,7 +59,8 @@ import javax.microedition.khronos.opengles.GL10;
  * match what they expect.
  */
 @TargetApi(15)
-public final class SphericalSurfaceView extends GLSurfaceView {
+public final class SphericalSurfaceView extends GLSurfaceView
+    implements VideoFrameMetadataListener {
 
   /**
    * This listener can be used to be notified when the {@link Surface} associated with this view is
@@ -69,17 +75,6 @@ public final class SphericalSurfaceView extends GLSurfaceView {
      */
     void surfaceChanged(@Nullable Surface surface);
   }
-
-  // A spherical mesh for video should be large enough that there are no stereo artifacts.
-  private static final int SPHERE_RADIUS_METERS = 50;
-
-  // TODO These should be configured based on the video type. It's assumed 360 video here.
-  private static final int DEFAULT_SPHERE_HORIZONTAL_DEGREES = 360;
-  private static final int DEFAULT_SPHERE_VERTICAL_DEGREES = 180;
-
-  // The 360 x 180 sphere has 5 degree quads. Increase these if lines in videos look wavy.
-  private static final int DEFAULT_SPHERE_COLUMNS = 72;
-  private static final int DEFAULT_SPHERE_ROWS = 36;
 
   // Arbitrary vertical field of view.
   private static final int FIELD_OF_VIEW_DEGREES = 90;
@@ -99,6 +94,9 @@ public final class SphericalSurfaceView extends GLSurfaceView {
   private @Nullable SurfaceListener surfaceListener;
   private @Nullable SurfaceTexture surfaceTexture;
   private @Nullable Surface surface;
+  private @C.StreamType int defaultStereoMode;
+  private @C.StreamType int currentStereoMode;
+  private @Nullable byte[] currentProjectionData;
 
   public SphericalSurfaceView(Context context) {
     this(context, null);
@@ -107,6 +105,8 @@ public final class SphericalSurfaceView extends GLSurfaceView {
   public SphericalSurfaceView(Context context, @Nullable AttributeSet attributeSet) {
     super(context, attributeSet);
 
+    defaultStereoMode = C.STEREO_MODE_MONO;
+    currentStereoMode = C.STEREO_MODE_MONO;
     mainHandler = new Handler(Looper.getMainLooper());
 
     // Configure sensors and touch.
@@ -129,30 +129,16 @@ public final class SphericalSurfaceView extends GLSurfaceView {
     setEGLContextClientVersion(2);
     setRenderer(renderer);
     setOnTouchListener(touchTracker);
-
-    setStereoMode(C.STEREO_MODE_MONO);
   }
 
   /**
-   * Sets stereo mode of the media to be played.
+   * Sets the default stereo mode. If the played video doesn't contain a stereo mode the default one
+   * is used.
    *
-   * @param stereoMode One of {@link C#STEREO_MODE_MONO}, {@link C#STEREO_MODE_TOP_BOTTOM}, {@link
-   *     C#STEREO_MODE_LEFT_RIGHT}.
+   * @param stereoMode A {@link C.StereoMode} value.
    */
-  public void setStereoMode(@C.StereoMode int stereoMode) {
-    Assertions.checkState(
-        stereoMode == C.STEREO_MODE_MONO
-            || stereoMode == C.STEREO_MODE_TOP_BOTTOM
-            || stereoMode == C.STEREO_MODE_LEFT_RIGHT);
-    Mesh mesh =
-        Mesh.createUvSphere(
-            SPHERE_RADIUS_METERS,
-            DEFAULT_SPHERE_ROWS,
-            DEFAULT_SPHERE_COLUMNS,
-            DEFAULT_SPHERE_VERTICAL_DEGREES,
-            DEFAULT_SPHERE_HORIZONTAL_DEGREES,
-            stereoMode);
-    queueEvent(() -> renderer.scene.setMesh(mesh));
+  public void setDefaultStereoMode(@C.StereoMode int stereoMode) {
+    defaultStereoMode = stereoMode;
   }
 
   /** Returns the {@link Surface} associated with this view. */
@@ -167,6 +153,12 @@ public final class SphericalSurfaceView extends GLSurfaceView {
    */
   public void setSurfaceListener(@Nullable SurfaceListener listener) {
     surfaceListener = listener;
+  }
+
+  @Override
+  public void onVideoFrameAboutToBeRendered(
+      long presentationTimeUs, long releaseTimeNs, Format format) {
+    setProjection(format.projectionData, format.stereoMode, releaseTimeNs);
   }
 
   @Override
@@ -228,6 +220,35 @@ public final class SphericalSurfaceView extends GLSurfaceView {
     if (oldSurface != null) {
       oldSurface.release();
     }
+  }
+
+  /**
+   * Sets projection data and stereo mode of the media to be played.
+   *
+   * @param projectionData Contains the projection data to be rendered.
+   * @param stereoMode A {@link C.StereoMode} value.
+   * @param timeNs When then new projection should be used.
+   */
+  private void setProjection(
+      @Nullable byte[] projectionData, @C.StereoMode int stereoMode, long timeNs) {
+    byte[] oldProjectionData = currentProjectionData;
+    int oldStereoMode = currentStereoMode;
+    currentProjectionData = projectionData;
+    currentStereoMode = stereoMode == Format.NO_VALUE ? defaultStereoMode : stereoMode;
+    if (oldStereoMode == currentStereoMode
+        && Arrays.equals(oldProjectionData, currentProjectionData)) {
+      return;
+    }
+
+    Projection projectionFromData = null;
+    if (currentProjectionData != null) {
+      projectionFromData = ProjectionDecoder.decode(currentProjectionData, currentStereoMode);
+    }
+    Projection projection =
+        projectionFromData != null && ProjectionRenderer.isSupported(projectionFromData)
+            ? projectionFromData
+            : Projection.createEquirectangular(currentStereoMode);
+    queueEvent(() -> renderer.scene.setProjection(projection, timeNs));
   }
 
   /** Detects sensor events and saves them as a matrix. */
@@ -328,7 +349,7 @@ public final class SphericalSurfaceView extends GLSurfaceView {
     private final float[] tempMatrix = new float[16];
 
     public Renderer() {
-      scene = new SceneRenderer();
+      scene = new SceneRenderer(Projection.createEquirectangular(C.STEREO_MODE_MONO));
       Matrix.setIdentityM(deviceOrientationMatrix, 0);
       Matrix.setIdentityM(touchPitchMatrix, 0);
       Matrix.setIdentityM(touchYawMatrix, 0);